Mako (absolute) Voltage control. (THX morfic)
This started out as just including the existing work of someone else.
But that was so obfuscated and needlessly moved in sysfs, i turned it
into the type control
we have on NP, meaning same sysfs location and output formatting.
Since the various kernels range from no preUV to different levels of UV
it is best to encourage the user
to think in absolute voltages.
"I can UV -150 on kernel X! Really? Odd, i can only UV kernel Y -50mV!"
Kernel X has no preUV, kernel Y has 100mV preUV, so they both run the
same.
This is needlessly confusing to users.

Talking in absolutes, users will get a better understanding of "what the
other guy's" device is capable of.
Sticking to a common sysfs location and to the common absolute voltage
is thus the right thing to do.

This seems to work now as it should.

This is heavily based on coolbho3k's NP patch. from Paul's stuff we
might have a few function names, but not the contents.
This is not his, if he claims otherwise, he is wrong.

Conflicts:
	drivers/cpufreq/cpufreq.c